+++ docs/conf.py
@@ -37,6 +37,7 @@ extensions = [
     'sphinx.ext.doctest',
     'sphinx.ext.intersphinx',
     'sphinx.ext.viewcode',
+    'parverdoc',
 ]
 
 # Add any paths that contain templates here, relative to this directory.